backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpcore
    #   openai
    #   starlette
    #   watchfiles
bcrypt==4.0.1
    # via -r requirements.in
beanie==1.23.0
    # via -r requirements.in
black==23.9.1
    # via -r requirements.in
blinker==1.7.0
    # via flask
brotli==1.1.0
    # via geventhttpclient
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via
    #   -r requirements.in
    #   geventhttpclient
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   beanie
    #   black
    #   flask
    #   pip-tools
    #   uvicorn
configargparse==1.7
    # via locust
distro==1.9.0
    # via openai
dnspython==2.4.2
    # via
    #   -r requirements.in
    #   email-validator
    #   pymongo
ecdsa==0.18.0
    # via python-jose
email-validator==2.0.0.post2
    # via fastapi
fastapi[all]==0.103.2
    # via -r requirements.in
flake8==6.1.0
    # via -r requirements.in
flask==3.0.2
    # via
    #   flask-cors
    #   flask-login
    #   locust
flask-cors==4.0.0
    # via locust
flask-login==0.6.3
    # via locust
gevent==23.9.1
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==2.0.11
    # via locust
greenlet==3.0.3
    # via gevent
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.18.0
    # via httpx
httptools==0.6.0
    # via uvicorn
httpx==0.25.0
    # via
    #   fastapi
    #   openai
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.0.0
    # via pytest
itsdangerous==2.1.2
    # via
    #   fastapi
    #   flask
jinja2==3.1.2
    # via
    #   fastapi
    #   flask
lazy-model==0.2.0
    # via beanie
locust==2.22.0
    # via -r requirements.in
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
motor==3.3.1
    # via
    #   -r requirements.in
    #   beanie
msgpack==1.0.7
    # via locust
mypy-extensions==1.0.0
    # via black
openai==1.13.3
    # via -r requirements.in
orjson==3.9.8
    # via fastapi
packaging==23.2
    # via
    #   black
    #   build
    #   pytest
passlib==1.7.4
    # via -r requirements.in
pathspec==0.11.2
    # via black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==3.11.0
    # via black
pluggy==1.4.0
    # via pytest
psutil==5.9.8
    # via locust
pyasn1==0.5.0
    # via
    #   python-jose
    #   rsa
pycodestyle==2.11.0
    # via flake8
pydantic==2.4.2
    # via
    #   -r requirements.in
    #   beanie
    #   fastapi
    #   lazy-model
    #   openai
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-extra-types==2.1.0
    # via fastapi
pydantic-settings==2.0.3
    # via
    #   -r requirements.in
    #   fastapi
pyflakes==3.1.0
    # via flake8
pymongo==4.5.0
    # via motor
pyproject-hooks==1.0.0
    # via build
pytest==8.0.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.6
    # via fastapi
pyyaml==6.0.1
    # via
    #   fastapi
    #   uvicorn
pyzmq==25.1.2
    # via locust
requests==2.31.0
    # via
    #   -r requirements.in
    #   locust
roundrobin==0.0.4
    # via locust
rsa==4.9
    # via python-jose
six==1.16.0
    # via
    #   ecdsa
    #   geventhttpclient
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
    #   openai
starlette==0.27.0
    # via fastapi
toml==0.10.2
    # via beanie
tqdm==4.66.2
    # via openai
typing-extensions==4.8.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
ujson==5.8.0
    # via fastapi
urllib3==2.2.0
    # via requests
uvicorn[standard]==0.23.2
    # via
    #   -r requirements.in
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.20.0
    # via uvicorn
websockets==11.0.3
    # via uvicorn
werkzeug==3.0.1
    # via
    #   flask
    #   flask-login
    #   locust
wheel==0.41.2
    # via pip-tools
zope-event==5.0
    # via gevent
zope-interface==6.1
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
